Rename, add Command
